WIP: code is now compilable
Code is now compilable. Though still lack a lot of components.
Only 2 modules are now compiled:
  - s5common, common library for both server and client side
  - s5afs, S5 All Flash System server, the server side

+++ Common/include/s5conf.h
@@ -11,6 +11,7 @@
  *
  * @author xxx
  */
+#include "basetype.h"
 
 typedef void* conf_file_t;		///< configuration object
 
@@ -168,7 +169,7 @@ int conf_get_int(conf_file_t conf , const char* section, const char* key, int de
  * @param[in]	    fatal_on_error	whether to LOG_FATAL if config not found
  * @returns		value of configuration item.
  */
-int conf_get_double(conf_file_t conf, const char* section, const char* key, double def_val, BOOL fatal_on_error);
+double conf_get_double(conf_file_t conf, const char* section, const char* key, double def_val, BOOL fatal_on_error);
 
 
 #ifdef __cplusplus
